fix #10 support boolean result from lit
lit has added a boolean as result to the updateComplete
promise. The value of the boolean returns false if
other updates are scheduled, and true if this was the
last event in the queue.

We can either approach that naively, and await the
next element, or we loop through all of them. For now,
we start with the naive approach and develop from
there.